Handled leading slash in path sending to http2
When urls returned from Apsis queueing system, the path will include a
leading slash.

This conflicts with http2, as it always prefixes the specified urls with
slash, and this ends up sending double slashes in the urls - which leads
into the classical "A fatal internal error occured, please contact
support" exception.