Disable http2-raise_errors to expose Apsis errors
As Apsis sends back "500 Internal Server Error" in cases like "Failed because there is already a subscription on this mailinglist for this subscriber", this message won't get lost due to not passing it on.

By disabling `raise_errors` making `http2` not running `handle_errors`, will pass all http errors as back to the project using the gem as expected, making it possible to take action - instead of the more generic and rather anonymous "500 Internal Server Error".

As this has worked as expected before, and neither `http2` nor `apsis-on-steroids` seems to have changes in this areas, it could seem like Apsis API has started returning different http-codes than previously.